+++ src/oersted/constants.py
@@ -1,0 +1,3 @@
+from numpy import pi
+
+MU0: float = 4.0 * pi * 10**-7