+++ CHANGELOG.md
@@ -43,3 +43,4 @@
 * Add function for calculating surface face areas and normals
 * Add Maxwell stress tensor and surface force calculation functions
 * Add function for creating tetradhedrons at surface faces for magnetization calc
+* Added surface force test